Split more CLI commands
